refactor(storage): extract SubjectBarrierMixin from governance (SQLite + ABC)
Verbatim move of the SubjectBarrier bucket (5 public methods + 7 privates,
the GDPR subject-erasure-barrier lifecycle) out of the residual governance
mixin into a new SubjectBarrierMixin, mirroring the Audit/Purge extraction.

- New `governance/_subject_barrier.py` in sqlite_storage and storage_base;
  re-exported from each governance package and composed into SQLiteStorage /
  BaseStorage alongside AuditEventStoreMixin / PurgeOperationStoreMixin.
- `complete_subject_erasure_barrier_after_empty_check` (one of the two
  successful-ERASE audit writers) moved byte-for-byte, including the
  ERASE-audit-idempotency block and the rowcount!=1-guarded barrier flip;
  `_append_audit_event_with_cursor` and `get_purge_operation` resolve via MRO.
- Residual keeps the shared module-level helpers (`_row_to_*`) which the new
  mixin imports; ruff dropped the now-unused governance.config / error /
  validator imports.
- Repoint `test_governance_refs.py` to assert the barrier methods are abstract
  on SubjectBarrierMixin (they no longer live on the residual GovernanceMixin).
