+++ reflexio/server/services/storage/sqlite_storage/governance/__init__.py
@@ -1,7 +1,9 @@
 from ._audit import AuditEventStoreMixin
 from ._purge import PurgeOperationStoreMixin
+from ._subject_barrier import SubjectBarrierMixin
 
 __all__ = [
     "AuditEventStoreMixin",
     "PurgeOperationStoreMixin",
+    "SubjectBarrierMixin",
 ]
