+++ reflexio/server/services/storage/storage_base/__init__.py
@@ -31,7 +31,11 @@ from ._retrieval_log import RetrievalLogMixin
 from ._shadow_verdicts import ShadowVerdictsMixin
 from ._share_links import ShareLinkMixin
 from ._stall_state import StallStateMixin
-from .governance import AuditEventStoreMixin, PurgeOperationStoreMixin
+from .governance import (
+    AuditEventStoreMixin,
+    PurgeOperationStoreMixin,
+    SubjectBarrierMixin,
+)
 from .playbook import (
     AgentEvaluationResultStoreMixin,
     AgentPlaybookStoreMixin,
@@ -56,6 +60,7 @@ class BaseStorage(
     RetrievalLogMixin,
     AuditEventStoreMixin,
     PurgeOperationStoreMixin,
+    SubjectBarrierMixin,
     GovernanceMixin,
     LineageEventMixin,
     OperationMixin,
@@ -269,6 +274,7 @@ __all__ = [
     "AgentEvaluationResultStoreMixin",
     "AuditEventStoreMixin",
     "PurgeOperationStoreMixin",
+    "SubjectBarrierMixin",
     "AgentPlaybookStoreMixin",
     "OptimizationJobStoreMixin",
     "PlaybookSourceLinkageMixin",
